requirements.txt
elasticsearch==7.0.1
flask-restful==0.3.7
scikit-criteria==0.2.10
